refactor: moved crawlers from src to root docs: updated readme to run crawlers
