+++ README.md
@@ -37,4 +37,10 @@ chmod +x keyfile.sh
 `
 
 ### Start Docker 
-docker compose up
+`docker compose up`
+
+### Run Crawlers
+`
+ docker exec -it academia-backend-server-1 bash
+ node crawlers/{COLLEGE-NAME}/{DEPARTMENT}.js
+ `
