+++ crawlers/IIT_Madras/cse.js
@@ -1,9 +1,9 @@
 const axios = require('axios');
 const cheerio = require('cheerio');
-const collegeService = require('../../modules/colleges/services');
-const professorsService = require('../../modules/professors/services');
+const collegeService = require('../../src/modules/colleges/services');
+const professorsService = require('../../src/modules/professors/services');
 const utils = require('../utils');
-const logger = require('../../config/logger');
+const logger = require('../../src/config/logger');
 
 const URL = 'https://www.cse.iitm.ac.in/listpeople.php?arg=MSQw';
 const BASE_URL = 'https://www.cse.iitm.ac.in/';
@@ -89,7 +89,7 @@ async function scrapeProfilesCSE() {
       const data = await fetchProfile(link);
       if (data) results.push(data);
     }
-
+    logger.info(`Number of professors: ${results.length}`);
     return results;
   } catch (err) {
     logger.error('Error fetching page:', err.message);
