+++ crawlers/IIT_Madras/mech.js
@@ -1,9 +1,9 @@
 const axios = require('axios');
 const cheerio = require('cheerio');
-const collegeService = require('../../modules/colleges/services');
-const professorsService = require('../../modules/professors/services');
+const collegeService = require('../../src/modules/colleges/services');
+const professorsService = require('../../src/modules/professors/services');
 const utils = require('../utils');
-const logger = require('../../config/logger');
+const logger = require('../../src/config/logger');
 
 const baseUrl = 'https://mech.iitm.ac.in/';
 const targetUrl = baseUrl + 'faculty.php';
@@ -117,7 +117,7 @@ async function scrapeProfilesMech() {
       const profile = await scrapeProfile(link);
       if (profile) results.push(profile);
     }
-
+    logger.info(`Number of professors: ${results.length}`);
     return results;
   } catch (err) {
     logger.error('Error fetching page:', err.message);
